preinit blockchain events
